cli: implement login and register operations
First, the argument parser was cli implemented.

Login and Register set the env var VEGETATION_TOKEN. To make it
permanent I took advantage of set_env crate which sets it to the profile
files for unix systems or to registry for Windows.

TODO: just realized that my fix for current process is not really a fix.
After login/register I still need to restart my shell because
env::set_var will not update it for current shell instance but only for
running process....